// src/app/models/TableOrderData.ts
export interface ITableOrderData {
  id: string;
  status: string;
  type: string;
  rawType: string;
  rawAmount: string;
  locality: string;
  division: string;
  deliveryType: string;
  deliveryAddress: string;
  hasAssistant: string;
  customerCarNumber: string;
  customerOrganizationLegalName: string;
  customerOrganizationActualName: string;
  customerContactName: string;
  customerContactPhone: string;
  paymentMethod: string;
  paymentMethodData: string;
  manager: string;
  driver: string;
  weigher: string;
  customerComment: string;
  employeeComment: string;
  desiredPickupDate: string;
  createdAt: string;
  weighedRaw: string;
  weighedRawAmount: string;
}